pyproject.toml
[tool.poetry]
name = "user-service"
version = "0.1.0"
description = ""
authors = ["A6e <A6E@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10.7"
aiomysql = "^0.1.1"
cattrs = "^22.2.0"
black = "^22.10.0"
isort = "^5.10.1"
orjson = "^3.8.1"
bcrypt = "^4.0.1"
aiohttp = "^3.8.3"
aiohttp_cors = "^0.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
